A first prototype of the 'handel_mcs' driver has been built on Linux.  Not yet tested.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@4192 42547f41-83fc-0310-9aef-9668e019ab9e

+++ modules/xia_handel_map/d_handel_mcs.h
@@ -26,8 +26,8 @@ typedef struct {
 	unsigned long handel_mcs_flags;
 
 	unsigned long buffer_length;
-	unsigned long *buffer_a;
-	unsigned long *buffer_b;
+	uint16_t *buffer_a;
+	uint16_t *buffer_b;
 } MX_HANDEL_MCS;
 
 /* Define all of the interface functions. */
